+++ app/models/email_address.rb
@@ -1,5 +1,6 @@
 class EmailAddress < ApplicationRecord
   include Concerns::BeforeCreateSetValidValuesBasedOnSource
+  include Concerns::AfterValidationSetSourceToMPDX
   include HasPrimary
   @@primary_scope = :person